chore:(refactor) to move more code into a git/helm specific file
