+++ pkg/jx/cmd/common.go
@@ -3,17 +3,14 @@ package cmd
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/url"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
 	"github.com/jenkins-x/golang-jenkins"
 	"github.com/jenkins-x/jx/pkg/auth"
 	"github.com/jenkins-x/jx/pkg/client/clientset/versioned"
-	"github.com/jenkins-x/jx/pkg/gits"
 	"github.com/jenkins-x/jx/pkg/jx/cmd/table"
 	cmdutil "github.com/jenkins-x/jx/pkg/jx/cmd/util"
 	"github.com/jenkins-x/jx/pkg/kube"
@@ -56,12 +53,6 @@ type ServerFlags struct {
 
 func (f *ServerFlags) IsEmpty() bool {
 	return f.ServerName == "" && f.ServerURL == ""
-}
-
-func addGitRepoOptionsArguments(cmd *cobra.Command, repositoryOptions *gits.GitRepositoryOptions) {
-	cmd.Flags().StringVarP(&repositoryOptions.ServerURL, "git-provider-url", "", "", "The git server URL to create new git repositories inside")
-	cmd.Flags().StringVarP(&repositoryOptions.Username, "git-username", "", "", "The git username to use for creating new git repositories")
-	cmd.Flags().StringVarP(&repositoryOptions.ApiToken, "git-api-token", "", "", "The git API token to use for creating new git repositories")
 }
 
 func (c *CommonOptions) Stdout() io.Writer {
@@ -147,49 +138,6 @@ func (o *CommonOptions) JXClientAndDevNamespace() (*versioned.Clientset, string,
 	return o.jxClient, o.devNamespace, nil
 }
 
-func (o *CommonOptions) GitServerKind(gitInfo *gits.GitRepositoryInfo) (string, error) {
-	return o.GitServerHostURLKind(gitInfo.HostURL())
-}
-
-func (o *CommonOptions) GitServerHostURLKind(hostURL string) (string, error) {
-	jxClient, devNs, err := o.JXClientAndDevNamespace()
-	if err != nil {
-		return "", err
-	}
-
-	kubeClient, _, err := o.KubeClient()
-	if err != nil {
-		return "", err
-	}
-
-	apisClient, err := o.Factory.CreateApiExtensionsClient()
-	if err != nil {
-		return "", err
-	}
-	err = kube.RegisterGitServiceCRD(apisClient)
-	if err != nil {
-		return "", err
-	}
-
-	kind, err := kube.GetGitServiceKind(jxClient, kubeClient, devNs, hostURL)
-	if err != nil {
-		return kind, err
-	}
-	if kind == "" {
-		if o.BatchMode {
-			return "", fmt.Errorf("No git server kind could be found for URL %s\nPlease try specify it via: jx create git server someKind %s", hostURL, hostURL)
-		}
-		kind, err = util.PickName(gits.KindGits, "Pick what kind of git server is")
-		if err != nil {
-			return "", err
-		}
-		if kind == "" {
-			return "", fmt.Errorf("No git kind chosen!")
-		}
-	}
-	return kind, nil
-}
-
 func (o *CommonOptions) JenkinsClient() (*gojenkins.Jenkins, error) {
 	if o.jenkinsClient == nil {
 		jenkins, err := o.Factory.CreateJenkinsClient()
@@ -212,23 +160,6 @@ func (o *CommonOptions) TeamAndEnvironmentNames() (string, string, error) {
 // warnf generates a warning
 func (o *CommonOptions) warnf(format string, a ...interface{}) {
 	o.Printf(util.ColorWarning("WARNING: "+format), a...)
-}
-
-// gitProviderForURL returns a GitProvider for the given git URL
-func (o *CommonOptions) gitProviderForURL(gitURL string, message string) (gits.GitProvider, error) {
-	gitInfo, err := gits.ParseGitURL(gitURL)
-	if err != nil {
-		return nil, err
-	}
-	authConfigSvc, err := o.Factory.CreateGitAuthConfigService()
-	if err != nil {
-		return nil, err
-	}
-	gitKind, err := o.GitServerKind(gitInfo)
-	if err != nil {
-		return nil, err
-	}
-	return gitInfo.PickOrCreateProvider(authConfigSvc, message, o.BatchMode, gitKind)
 }
 
 func (o *ServerFlags) addGitServerFlags(cmd *cobra.Command) {
@@ -417,103 +348,6 @@ func (o *CommonOptions) findServiceInNamespace(name string, ns string) (string, 
 	return url, nil
 }
 
-func (o *CommonOptions) registerLocalHelmRepo(repoName, ns string) error {
-	if repoName == "" {
-		repoName = kube.LocalHelmRepoName
-	}
-	// TODO we should use the auth package to keep a list of server login/pwds
-	// TODO we have a chartmuseumAuth.yaml now but sure yet if that's the best thing to do
-	username := "admin"
-	password := "admin"
-
-	// lets check if we have a local helm repository
-	client, _, err := o.Factory.CreateClient()
-	if err != nil {
-		return err
-	}
-	u, err := kube.FindServiceURL(client, ns, kube.ServiceChartMuseum)
-	if err != nil {
-		return err
-	}
-	u2, err := url.Parse(u)
-	if err != nil {
-		return err
-	}
-	if u2.User == nil {
-		u2.User = url.UserPassword(username, password)
-	}
-	helmUrl := u2.String()
-	// lets check if we already have the helm repo installed or if we need to add it or remove + add it
-	text, err := o.getCommandOutput("", "helm", "repo", "list")
-	if err != nil {
-		return err
-	}
-	lines := strings.Split(text, "\n")
-	remove := false
-	for _, line := range lines {
-		t := strings.TrimSpace(line)
-		if t != "" {
-			fields := strings.Fields(t)
-			if len(fields) > 1 {
-				if fields[0] == repoName {
-					if fields[1] == helmUrl {
-						return nil
-					} else {
-						remove = true
-					}
-				}
-			}
-		}
-	}
-	if remove {
-		err = o.runCommand("helm", "repo", "remove", repoName)
-		if err != nil {
-			return err
-		}
-	}
-	return o.runCommand("helm", "repo", "add", repoName, helmUrl)
-}
-
-// installChart installs the given chart
-func (o *CommonOptions) installChart(releaseName string, chart string, version string, ns string, helmUpdate bool, setValues []string) error {
-	if helmUpdate {
-		err := o.runCommand("helm", "repo", "update")
-		if err != nil {
-			return err
-		}
-	}
-	timeout := fmt.Sprintf("--timeout=%s", defaultInstallTimeout)
-	args := []string{"upgrade", "--install", timeout}
-	if version != "" {
-		args = append(args, "--version", version)
-	}
-	if ns != "" {
-		kubeClient, _, err := o.KubeClient()
-		if err != nil {
-			return err
-		}
-		annotations := map[string]string{"jenkins-x.io/created-by": "Jenkins X"}
-		kube.EnsureNamespaceCreated(kubeClient, ns, nil, annotations)
-		args = append(args, "--namespace", ns)
-	}
-	for _, value := range setValues {
-		args = append(args, "--set", value)
-		o.Printf("Set chart value: --set %s\n", util.ColorInfo(value))
-	}
-	args = append(args, releaseName, chart)
-	return o.runCommand("helm", args...)
-}
-
-// deleteChart deletes the given chart
-func (o *CommonOptions) deleteChart(releaseName string, purge bool) error {
-	args := []string{"delete"}
-	if purge {
-		args = append(args, "--purge")
-	}
-	args = append(args, releaseName)
-	return o.runCommand("helm", args...)
-}
-
 func (o *CommonOptions) retry(attempts int, sleep time.Duration, call func() error) (err error) {
 	for i := 0; ; i++ {
 		err = call()
@@ -642,70 +476,6 @@ func (o *CommonOptions) pickRemoteURL(config *gitcfg.Config) (string, error) {
 		err := survey.AskOne(prompt, &url, nil)
 		if err != nil {
 			return "", err
-		}
-	}
-	return url, nil
-}
-
-func (*CommonOptions) FindHelmChart() (string, error) {
-	dir, err := os.Getwd()
-	if err != nil {
-		return "", err
-	}
-	// lets try find the chart file
-	chartFile := filepath.Join(dir, "Chart.yaml")
-	exists, err := util.FileExists(chartFile)
-	if err != nil {
-		return "", err
-	}
-	if !exists {
-		// lets try find all the chart files
-		files, err := filepath.Glob("*/Chart.yaml")
-		if err != nil {
-			return "", err
-		}
-		if len(files) > 0 {
-			chartFile = files[0]
-		} else {
-			files, err = filepath.Glob("*/*/Chart.yaml")
-			if err != nil {
-				return "", err
-			}
-			if len(files) > 0 {
-				chartFile = files[0]
-				return chartFile, nil
-			}
-		}
-	}
-	return "", nil
-}
-
-func (o *CommonOptions) discoverGitURL(gitConf string) (string, error) {
-	if gitConf == "" {
-		return "", fmt.Errorf("No GitConfDir defined!")
-	}
-	cfg := gitcfg.NewConfig()
-	data, err := ioutil.ReadFile(gitConf)
-	if err != nil {
-		return "", fmt.Errorf("Failed to load %s due to %s", gitConf, err)
-	}
-
-	err = cfg.Unmarshal(data)
-	if err != nil {
-		return "", fmt.Errorf("Failed to unmarshal %s due to %s", gitConf, err)
-	}
-	remotes := cfg.Remotes
-	if len(remotes) == 0 {
-		return "", nil
-	}
-	url := gits.GetRemoteUrl(cfg, "origin")
-	if url == "" {
-		url = gits.GetRemoteUrl(cfg, "upstream")
-		if url == "" {
-			url, err = o.pickRemoteURL(cfg)
-			if err != nil {
-				return "", err
-			}
 		}
 	}
 	return url, nil
