+++ pkg/jx/cmd/common_git.go
@@ -2,11 +2,15 @@ package cmd
 
 import (
 	"fmt"
+	"io/ioutil"
 
 	"github.com/jenkins-x/jx/pkg/auth"
 	"github.com/jenkins-x/jx/pkg/gits"
 	"github.com/jenkins-x/jx/pkg/issues"
 	"github.com/jenkins-x/jx/pkg/kube"
+	"github.com/jenkins-x/jx/pkg/util"
+	"github.com/spf13/cobra"
+	gitcfg "gopkg.in/src-d/go-git.v4/config"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -125,4 +129,101 @@ func (o *CommonOptions) ensureGitServiceCRD(server *auth.AuthServer) error {
 		return err
 	}
 	return kube.EnsureGitServiceExistsForHost(jxClient, devNs, kind, server.Name, server.URL, o.Out)
+}
+
+func (o *CommonOptions) discoverGitURL(gitConf string) (string, error) {
+	if gitConf == "" {
+		return "", fmt.Errorf("No GitConfDir defined!")
+	}
+	cfg := gitcfg.NewConfig()
+	data, err := ioutil.ReadFile(gitConf)
+	if err != nil {
+		return "", fmt.Errorf("Failed to load %s due to %s", gitConf, err)
+	}
+
+	err = cfg.Unmarshal(data)
+	if err != nil {
+		return "", fmt.Errorf("Failed to unmarshal %s due to %s", gitConf, err)
+	}
+	remotes := cfg.Remotes
+	if len(remotes) == 0 {
+		return "", nil
+	}
+	url := gits.GetRemoteUrl(cfg, "origin")
+	if url == "" {
+		url = gits.GetRemoteUrl(cfg, "upstream")
+		if url == "" {
+			url, err = o.pickRemoteURL(cfg)
+			if err != nil {
+				return "", err
+			}
+		}
+	}
+	return url, nil
+}
+
+func addGitRepoOptionsArguments(cmd *cobra.Command, repositoryOptions *gits.GitRepositoryOptions) {
+	cmd.Flags().StringVarP(&repositoryOptions.ServerURL, "git-provider-url", "", "", "The git server URL to create new git repositories inside")
+	cmd.Flags().StringVarP(&repositoryOptions.Username, "git-username", "", "", "The git username to use for creating new git repositories")
+	cmd.Flags().StringVarP(&repositoryOptions.ApiToken, "git-api-token", "", "", "The git API token to use for creating new git repositories")
+}
+
+func (o *CommonOptions) GitServerKind(gitInfo *gits.GitRepositoryInfo) (string, error) {
+	return o.GitServerHostURLKind(gitInfo.HostURL())
+}
+
+func (o *CommonOptions) GitServerHostURLKind(hostURL string) (string, error) {
+	jxClient, devNs, err := o.JXClientAndDevNamespace()
+	if err != nil {
+		return "", err
+	}
+
+	kubeClient, _, err := o.KubeClient()
+	if err != nil {
+		return "", err
+	}
+
+	apisClient, err := o.Factory.CreateApiExtensionsClient()
+	if err != nil {
+		return "", err
+	}
+	err = kube.RegisterGitServiceCRD(apisClient)
+	if err != nil {
+		return "", err
+	}
+
+	kind, err := kube.GetGitServiceKind(jxClient, kubeClient, devNs, hostURL)
+	if err != nil {
+		return kind, err
+	}
+	if kind == "" {
+		if o.BatchMode {
+			return "", fmt.Errorf("No git server kind could be found for URL %s\nPlease try specify it via: jx create git server someKind %s", hostURL, hostURL)
+		}
+		kind, err = util.PickName(gits.KindGits, "Pick what kind of git server is")
+		if err != nil {
+			return "", err
+		}
+		if kind == "" {
+			return "", fmt.Errorf("No git kind chosen!")
+		}
+	}
+	return kind, nil
+}
+
+// gitProviderForURL returns a GitProvider for the given git URL
+func (o *CommonOptions) gitProviderForURL(gitURL string, message string) (gits.GitProvider, error) {
+	gitInfo, err := gits.ParseGitURL(gitURL)
+	if err != nil {
+		return nil, err
+	}
+	authConfigSvc, err := o.Factory.CreateGitAuthConfigService()
+	if err != nil {
+		return nil, err
+	}
+	gitKind, err := o.GitServerKind(gitInfo)
+	if err != nil {
+		return nil, err
+	}
+	return gitInfo.PickOrCreateProvider(authConfigSvc, message, o.BatchMode, gitKind)
 }
